third_party/WebKit/LayoutTests/external/wpt/FileAPI/url/cross-global-revoke.sub-expected.txt

This is a testharness.js-based test.
FAIL It is possible to revoke same-origin blob URLs from different frames. assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL It is possible to revoke same-origin blob URLs from a different worker global. assert_unreached: Should have rejected: undefined Reached unreachable code
PASS It is not possible to revoke cross-origin blob URLs.
Harness: the test ran to completion.
